+++ README.md
@@ -1,24 +1,24 @@
 # Introduction
 
-  - FlowMetrics helps visualize GitHub Actions metrics over time, providing insights into your workflow trends. By connecting to GitHub's REST API, FlowMetrics extends the storage of your workflow metrics beyond GitHub's 90-day retention period.
+  - CICDash helps visualize GitHub Actions metrics over time, providing insights into your workflow trends. By connecting to GitHub's REST API, CICDash extends the storage of your workflow metrics beyond GitHub's 90-day retention period.
 By storing a history of your GitHub workflow metrics, trends can be more easily identified, saving time locating and debugging any issues that appear throughout development.
 
   - Workflow Runs by Month, Lifetime Workflow Attempts, Average Run Time: Monthly vs Lifetime (seconds) - Two different graphs for this one, Execution Time (seconds), Failure Rate By Job Step
 
 
-  - For more information, please reference our website at https://cicdmetrics.com and medium article "FlowMetrics: A GitHub Actions Metrics Dashboard"
+  - For more information, please reference our website at https://cicdmetrics.com and medium article "CICDash: An Open Source GitHub Actions Metrics Dashboard"
 
 ## Getting started
-  - FlowMetrics requires a working GitHub Actions (GHA) enabled repository.
+  - CICDash requires a working GitHub Actions (GHA) enabled repository.
   - If you need help setting that up, please refer to GitHub's documentation: https://docs.github.com/en/actions/learn-github-actions/understanding-github-actions
-Otherwise, please proceed with the following steps to visualize GitHub workflow data with FlowMetrics.
+Otherwise, please proceed with the following steps to visualize GitHub workflow data with CICDash.
 
 
-## Install FlowMetrics
+## Install CICDash
 
 
-  - Ensure your repository is enabled for GitHub Actions. If not, please enable GitHub Actions in your repository settings. 
-Install FlowMetrics from GitHub Marketplace onto your GHA enabled repo.
+  - Ensure your repository is enabled for GitHub Actions. If not, please enable GitHub Actions in your repository settings.
+Install CICDash from GitHub Marketplace onto your GHA enabled repo.
 
 ## Login with Github
 
@@ -33,8 +33,8 @@ Install FlowMetrics from GitHub Marketplace onto your GHA enabled repo.
 
 # Contribution requirements
 
-Contributions are essential in the open-source community. If you would like to contribute make FlowMetrics even better together, please follow the steps below:
-1.   Fork the project: 
+Contributions are essential in the open-source community. If you would like to contribute make CICDash even better together, please follow the steps below:
+1.   Fork the project:
 2.   Create a feature branch for your additions
 3.   Submit a pull request that details your changes.
 4.   We'll review your feature(s) and get back to you.
