+++ frontend/public/index.html
@@ -11,7 +11,7 @@
         margin: 0;
       }
     </style>
-    <title>FlowMetrics</title>
+    <title>CICDash</title>
   </head>
 
   <body>
